#commit 11 - refactoring of CryptoExchangeService
