+++ MetaExchanger/MetaExchanger.Application/Services/ICryptoExchangeService.cs
@@ -3,8 +3,11 @@ using MetaExchanger.Application.Domain;
 
 namespace MetaExchanger.Application.Services
 {
+    /// <summary>
+    /// Interface with abstract definitions of crypto exchange service.
+    /// </summary>
     public interface ICryptoExchangeService
     {
-        Task<Result<IEnumerable<DomainOrder>>> GetOrdersAsync(DomainOrder movie, CancellationToken token = default);
+        Task<Result<IEnumerable<DomainOrder>>> GetBestExecutionAsync(DomainOrder movie, CancellationToken token = default);
     }
 }
